+++ t-fusionTool/src/main/java/eu/unifiedviews/plugins/transformer/fusiontool/io/DataUnitRDFWriterWithMetadata.java
@@ -5,12 +5,12 @@ import cz.cuni.mff.odcleanstore.fusiontool.writers.CloseableRDFWriterBase;
 import cz.cuni.mff.odcleanstore.vocabulary.ODCS;
 import eu.unifiedviews.dataunit.DataUnitException;
 import eu.unifiedviews.dataunit.rdf.WritableRDFDataUnit;
-import org.openrdf.model.Resource;
-import org.openrdf.model.Statement;
-import org.openrdf.model.URI;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.repository.RepositoryConnection;
-import org.openrdf.repository.RepositoryException;
+import org.eclipse.rdf4j.model.Resource;
+import org.eclipse.rdf4j.model.Statement;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.ValueFactory;
+import org.eclipse.rdf4j.repository.RepositoryConnection;
+import org.eclipse.rdf4j.repository.RepositoryException;
 
 import java.io.IOException;
 import java.util.concurrent.atomic.AtomicLong;
@@ -18,10 +18,10 @@ import java.util.concurrent.atomic.AtomicLong;
 public class DataUnitRDFWriterWithMetadata extends CloseableRDFWriterBase {
     public static final String DATA_GRAPH_NAME_INFIX = "-";
     private final RepositoryConnection connection;
-    private final URI defaultContext;
+    private final IRI defaultContext;
     private final WritableRDFDataUnit dataUnit;
     private final String dataGraphSymbolicName;
-    private final URI metadataContext;
+    private final IRI metadataContext;
     private final AtomicLong counter = new AtomicLong(0);
     private final ValueFactory valueFactory;
 
@@ -46,7 +46,7 @@ public class DataUnitRDFWriterWithMetadata extends CloseableRDFWriterBase {
     @Override
     public void write(ResolvedStatement resolvedStatement) throws IOException {
         try {
-            URI statementContext = dataUnit.addNewDataGraph(dataGraphSymbolicName + DATA_GRAPH_NAME_INFIX + Long.toString(counter.incrementAndGet()));
+            IRI statementContext = dataUnit.addNewDataGraph(dataGraphSymbolicName + DATA_GRAPH_NAME_INFIX + Long.toString(counter.incrementAndGet()));
             connection.add(resolvedStatement.getStatement(), statementContext);
             connection.add(statementContext, ODCS.QUALITY, valueFactory.createLiteral(resolvedStatement.getQuality()), metadataContext);
             for (Resource sourceGraph : resolvedStatement.getSourceGraphNames()) {