+++ t-relationalToRdf/src/main/java/eu/unifiedviews/plugins/transformer/relationaltordf/RelationalToRdf.java
@@ -4,7 +4,7 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.List;
 
-import org.openrdf.model.URI;
+import org.eclipse.rdf4j.model.IRI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -107,11 +107,11 @@ public class RelationalToRdf extends AbstractDpu<RelationalToRdfConfig_V1> {
                         LOG.debug("Use table subject option is used, preparing subject for table");
                         // Prepare subject for table.
                         // TODO: We can use better subject here!
-                        final URI tableURI = this.faultTolerance.execute(new FaultTolerance.ActionReturn<URI>() {
+                        final IRI tableURI = this.faultTolerance.execute(new FaultTolerance.ActionReturn<IRI>() {
 
                             @Override
-                            public URI action() throws Exception {
-                                return RelationalToRdf.this.rdfTableWrap.getValueFactory().createURI(entry.getTableName());
+                            public IRI action() throws Exception {
+                                return RelationalToRdf.this.rdfTableWrap.getValueFactory().createIRI(entry.getTableName());
                             }
                         });
                         // Set as a table subject.