+++ l-rdfToVirtuoso/CHANGELOG.md
@@ -1,6 +1,10 @@
 L-RdfToVirtuoso
 ----------
 
+v2.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
 v1.0.3
 ---
 * Update to helpers 2.1.3