+++ t-gzipper/CHANGELOG.md
@@ -1,6 +1,10 @@
 T-Gzipper
 ----------
 
+v2.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
 v1.0.4
 ---
 * Improved slovak translations