+++ t-sparqlUpdate/CHANGELOG.md
@@ -1,11 +1,12 @@
 T-SparqlUpdate
 ----------
 
-v2.2.0
+v3.0.0
 ---
 * Support for [RDF Data Processing Performance Optimization](https://grips.semantic-web.at/display/UDDOC/RDF+Data+Processing+Performance+Optimization)
 * Connections properly closed
 * Fault tolerance support removed (not used anyway)
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
 
 v2.1.6
 ---