+++ t-tabular/CHANGELOG.md
@@ -1,6 +1,10 @@
 T-Tabular
 ----------
 
+v3.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
 v2.3.3
 ---
 * Added support for [RDF Validation extension](https://grips.semantic-web.at/display/UDDOC/RDF+Validation)